dispatch优化/重构
